requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
mysqlclient==2.2.0
alembic==1.12.1
pydantic-settings==2.0.3
python-dateutil==2.8.2
email-validator==2.1.0
cryptography==41.0.5
